+++ pepper/src/buffer_view.rs
@@ -37,7 +37,7 @@ pub struct BufferView {
     pub client_handle: ClientHandle,
     pub buffer_handle: BufferHandle,
     pub cursors: CursorCollection,
-    //pub(crate) scroll: BufferPositionIndex,
+    pub(crate) scroll: BufferPositionIndex,
 }
 
 impl BufferView {
@@ -599,6 +599,7 @@ impl BufferViewCollection {
             client_handle,
             buffer_handle,
             cursors: CursorCollection::new(),
+            scroll: 0,
         });
         handle
     }