+++ plugin-lsp/src/client_event_handler.rs
@@ -1101,4 +1101,3 @@ fn goto_definition(
         DefinitionLocation::Invalid => Ok(()),
     }
 }
-